ref api

+++ commerce/warehouse/src/main/java/ru/yandex/practicum/entity/ProductInWarehouse.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.entity;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -16,7 +15,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import ru.yandex.practicum.model.entity.product.Product;
+import ru.yandex.practicum.model.productEntity.Product;
 
 import java.util.UUID;
 
@@ -49,5 +48,5 @@ public class ProductInWarehouse {
     private Double weight;
 
     @Column(name = "available_stock", nullable = false)
-    private Integer availableStock = 0;
+    private Long availableStock = 0L;
 }
